Factor out a function for post content generation
